Fix Open Targets crash and supplement ClinicalTrials with drug-name searches
- models.py: Change KnownDrug.phase from int to float — Open Targets API
  returns fractional phases (e.g. 0.5), causing a Pydantic validation error
  that crashed the entire worker and zeroed out genetic score + tractability

- clinical_trials.py: Add supplement_with_drug_names() — gene-name searches
  miss trials filed under drug trade names (e.g. BACE1 → verubecestat,
  lanabecestat, atabecestat). Searches each known drug from Open Targets,
  deduplicates by NCT ID, and recomputes success rate.

- app.py: Call supplement_with_drug_names() after parallel workers complete,
  using the known drugs list returned by Open Targets.

Result for BACE1/Alzheimers: failed trials 2 → 10, genetic score 0.0 → 0.43,
tractability 0.0 → 0.8, recommendation NO-GO → CAUTION (scientifically correct).

+++ models.py
@@ -39,7 +39,7 @@ class GeneticAssociation(BaseModel):
 
 class KnownDrug(BaseModel):
     drug_name: str
-    phase: int
+    phase: float  # API can return fractional phases (e.g. 0.5 for Phase 1/2)
     status: str   # e.g. "Approved", "Phase 3", "Withdrawn"
     mechanism: str
